Find_Work coursework (third course)/Django/FIND_WORK/requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.0
cffi==1.14.3
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
fpdf==1.7.2
fpdf2==2.0.6
future==0.18.2
html5lib==1.1
numpy==1.19.4
Pillow==8.0.0
pkg-resources==0.0.0
psycopg2==2.8.6
pycparser==2.20
Pyphen==0.10.0
python-bidi==0.4.2
pytz==2020.4
reportlab==3.5.55
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
WeasyPrint==52.1
webencodings==0.5.1
wkhtmltopdf==0.2
xlwt==1.3.0
